Moved file_stream tests to test_responses.py
